[CNR-1027-B] Hide boolean storeErrors in repositories
