+++ core/src/main/java/org/hisp/dhis/android/core/arch/repositories/collection/ReadOnlyWithDownloadObjectRepository.java
@@ -38,16 +38,14 @@ public interface ReadOnlyWithDownloadObjectRepository<M extends CoreObject> exte
     /**
      * Downloads the resource in scope in an asynchronous way. As soon as it's downloaded and processed, the Completable
      * is completed.
-     * @param storeError whether to store errors in the database
      * @return a {@link Completable} that completes when the download and processing is finished
      */
-    Completable download(boolean storeError);
+    Completable download();
 
     /**
      * Downloads the resource in scope in a synchronous way. Important: this is a blocking method and it should not be
      * executed in the main thread. Consider the asynchronous version {@link #get()}. The method will finish
      * with a void as soon as the whole download and processing is finished.
-     * @param storeError whether to store errors in the database
      */
-    void blockingDownload(boolean storeError);
+    void blockingDownload();
 }
