+++ core/src/main/java/org/hisp/dhis/android/core/arch/repositories/object/internal/ReadOnlyFirstObjectWithDownloadRepositoryImpl.java
@@ -56,22 +56,20 @@ public class ReadOnlyFirstObjectWithDownloadRepositoryImpl<M extends CoreObject,
     /**
      * Downloads the resource in scope in an asynchronous way. As soon as it's downloaded and processed, the
      * {@link Completable} is completed.
-     * @param storeError whether to store errors in the database
      * @return a {@link Completable} that completes when the download and processing is finished
      */
     @Override
-    public Completable download(boolean storeError) {
-        return downloadCompletableProvider.getCompletable(storeError);
+    public Completable download() {
+        return downloadCompletableProvider.getCompletable(true);
     }
 
     /**
      * Downloads the resource in scope in a synchronous way. The method will finish
      * as soon as the whole download and processing is finished. Important: this is a blocking method and it should
-     * not be executed in the main thread. Consider the asynchronous version {@link #download}.
-     * @param storeError whether to store errors in the database
+     * not be executed in the main thread. Consider the asynchronous version {@link #download()}.
      */
     @Override
-    public void blockingDownload(boolean storeError) {
-        download(storeError).blockingAwait();
+    public void blockingDownload() {
+        download().blockingAwait();
     }
 }
