+++ core/src/main/java/org/hisp/dhis/android/core/domain/aggregated/data/internal/AggregatedDataCall.java
@@ -110,7 +110,7 @@ final class AggregatedDataCall {
     Observable<D2Progress> download() {
         D2ProgressManager progressManager = new D2ProgressManager(null);
 
-        Observable<D2Progress> observable = systemInfoRepository.download(true)
+        Observable<D2Progress> observable = systemInfoRepository.download()
                 .toSingle(() -> progressManager.increaseProgress(SystemInfo.class, false))
                 .flatMapObservable(progress -> selectDataSetsAndDownload(progressManager, progress));
         return rxCallExecutor.wrapObservableTransactionally(observable, true);
