+++ core/src/main/java/org/hisp/dhis/android/core/event/internal/EventWithLimitCallFactory.java
@@ -172,7 +172,7 @@ public final class EventWithLimitCallFactory {
     }
 
     private Observable<D2Progress> downloadSystemInfo(D2ProgressManager progressManager) {
-        return systemInfoRepository.download(true)
+        return systemInfoRepository.download()
                 .toSingle(() -> progressManager.increaseProgress(SystemInfo.class, false))
                 .toObservable();
     }
