+++ core/src/main/java/org/hisp/dhis/android/core/systeminfo/internal/SystemInfoEntityDIModule.java
@@ -26,27 +26,51 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.systeminfo;
+package org.hisp.dhis.android.core.systeminfo.internal;
 
-import org.hisp.dhis.android.core.systeminfo.internal.SystemInfoModuleImpl;
+import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter;
+import org.hisp.dhis.android.core.arch.db.stores.internal.ObjectWithoutUidStore;
+import org.hisp.dhis.android.core.arch.di.internal.ObjectWithoutUidEntityDIModule;
+import org.hisp.dhis.android.core.arch.handlers.internal.Handler;
+import org.hisp.dhis.android.core.arch.handlers.internal.ObjectWithoutUidHandlerImpl;
+import org.hisp.dhis.android.core.arch.repositories.children.internal.ChildrenAppender;
+import org.hisp.dhis.android.core.arch.repositories.collection.ReadOnlyWithDownloadObjectRepository;
+import org.hisp.dhis.android.core.systeminfo.SystemInfo;
+import org.hisp.dhis.android.core.systeminfo.SystemInfoObjectRepository;
+
+import java.util.Collections;
+import java.util.Map;
 
 import dagger.Module;
 import dagger.Provides;
 import dagger.Reusable;
-import retrofit2.Retrofit;
 
-@Module(includes = {SystemInfoEntityDIModule.class})
-public final class SystemInfoPackageDIModule {
+@Module
+public final class SystemInfoEntityDIModule implements ObjectWithoutUidEntityDIModule<SystemInfo> {
 
+    @Override
     @Provides
     @Reusable
-    SystemInfoService service(Retrofit retrofit) {
-        return retrofit.create(SystemInfoService.class);
+    public ObjectWithoutUidStore<SystemInfo> store(DatabaseAdapter databaseAdapter) {
+        return SystemInfoStore.create(databaseAdapter);
+    }
+
+    @Override
+    @Provides
+    @Reusable
+    public Handler<SystemInfo> handler(ObjectWithoutUidStore<SystemInfo> store) {
+        return new ObjectWithoutUidHandlerImpl<>(store);
     }
 
     @Provides
     @Reusable
-    SystemInfoModule systemInfoModule(SystemInfoModuleImpl impl) {
+    ReadOnlyWithDownloadObjectRepository<SystemInfo> systemInfoRepository(SystemInfoObjectRepository impl) {
         return impl;
+    }
+
+    @Provides
+    @Reusable
+    Map<String, ChildrenAppender<SystemInfo>> childrenAppenders() {
+        return Collections.emptyMap();
     }
 }
