+++ core/src/main/java/org/hisp/dhis/android/core/systeminfo/internal/SystemInfoModuleDownloader.java
@@ -48,6 +48,6 @@ public class SystemInfoModuleDownloader implements RxMetadataModuleDownloader {
 
     @Override
     public Completable downloadMetadata() {
-        return systemInfoRepository.download(true);
+        return systemInfoRepository.download();
     }
 }
