+++ core/src/main/java/org/hisp/dhis/android/core/trackedentity/internal/TrackedEntityInstanceWithLimitCallFactory.java
@@ -135,7 +135,7 @@ class TrackedEntityInstanceWithLimitCallFactory {
     }
 
     private Observable<D2Progress> downloadSystemInfo(D2ProgressManager progressManager) {
-        return systemInfoRepository.download(true)
+        return systemInfoRepository.download()
                 .toSingle(() -> progressManager.increaseProgress(SystemInfo.class, false))
                 .toObservable();
     }
