+++ core/src/main/java/org/hisp/dhis/android/core/user/internal/UserAuthenticateCallFactory.java
@@ -28,8 +28,6 @@
 
 package org.hisp.dhis.android.core.user.internal;
 
-import androidx.annotation.NonNull;
-
 import org.hisp.dhis.android.core.arch.api.executors.internal.APICallExecutor;
 import org.hisp.dhis.android.core.arch.api.internal.ServerUrlInterceptor;
 import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter;
@@ -55,6 +53,7 @@ import org.hisp.dhis.android.core.wipe.internal.WipeModule;
 
 import javax.inject.Inject;
 
+import androidx.annotation.NonNull;
 import dagger.Reusable;
 import io.reactivex.Single;
 import okhttp3.HttpUrl;
@@ -165,7 +164,7 @@ public final class UserAuthenticateCallFactory {
             authenticatedUserStore.updateOrInsertWhere(authenticatedUserToStore);
             credentialsSecureStore.setCredentials(Credentials.create(username, password));
 
-            systemInfoRepository.download(true).blockingAwait();
+            systemInfoRepository.download().blockingAwait();
 
             handleUser(authenticatedUser);
             transaction.setSuccessful();
